pyproject.toml
[project]
name = "fastapi-ai-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.135.1",
    "groq>=1.1.1",
    "httpx>=0.28.1",
    "langgraph>=1.1.2",
    "mcp>=1.26.0",
    "numpy>=2.4.3",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.9.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.1",
    "redis[asyncio]>=7.3.0",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "torch==2.2.2",
    "uvicorn[standard]>=0.42.0",
]
